Bigint

+++ backend/app/models/stock.py
@@ -1,5 +1,5 @@
 from sqlalchemy.orm import validates, relationship
-from sqlalchemy.types import Integer, String, Boolean
+from sqlalchemy.types import Integer, String, Boolean, BigInteger
 from .db import db, environment, SCHEMA, add_prefix_for_prod, watchlists_stocks
 from sqlalchemy.orm import declarative_mixin
 from .watchlist import Watchlist
@@ -13,7 +13,7 @@ class Stock(db.Model):
     if environment == "production":
         __table_args__ = {'schema': SCHEMA}
 
-    id = db.Column(db.Integer, primary_key=True)
+    id = db.Column(db.BigInteger, primary_key=True)
     stock_symbol = db.Column(db.String(5), nullable=False)
     company_name = db.Column(db.String(255), nullable=False)
